backend/requirements.txt

# FastAPI Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# LangChain & LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-groq>=0.1.3
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Processing
PyPDF2>=3.0.0
pandas>=2.0.0
numpy>=1.24.0

# Image & QR Processing
opencv-python>=4.8.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0

# Translation & Audio
deep-translator>=1.11.0
gTTS>=2.4.0
SpeechRecognition>=3.10.0
pydub>=0.25.1

# WebSocket
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
